+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- SDK should drop the event when any event processor returns nil [#1523](https://github.com/getsentry/sentry-ruby/pull/1523)
+
 ## 4.6.4
 
 - Extend Rake with a more elegant and reliable way [#1517](https://github.com/getsentry/sentry-ruby/pull/1517)